# frontend-web/src/app/api/model/userInput.ts
/**
 * Rekor API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserInput { 
    email: string;
    name: string;
}